fix: [hw9] correct spelling of notification sender service
