--- apps/notfication-sender/src/notfication-sender.module.ts
@@ -1,10 +1,0 @@
-import { Module } from '@nestjs/common';
-import { NotficationSenderController } from './notfication-sender.controller';
-import { NotficationSenderService } from './notfication-sender.service';
-
-@Module({
-  imports: [],
-  controllers: [NotficationSenderController],
-  providers: [NotficationSenderService],
-})
-export class NotficationSenderModule {}
