+++ apps/notification-sender/src/notification-sender.controller.ts
@@ -1,0 +1,12 @@
+import { Controller, Get } from '@nestjs/common';
+import { NotificationSenderService } from './notification-sender.service';
+
+@Controller()
+export class NotificationSenderController {
+  constructor(private readonly notificationSenderService: NotificationSenderService) {}
+
+  @Get()
+  getHello(): string {
+    return this.notificationSenderService.getHello();
+  }
+}
