+++ apps/notification-sender/src/notification-sender.module.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common';
+import { NotificationSenderController } from './notification-sender.controller';
+import { NotificationSenderService } from './notification-sender.service';
+
+@Module({
+  imports: [],
+  controllers: [NotificationSenderController],
+  providers: [NotificationSenderService],
+})
+export class NotificationSenderModule {}
