+++ apps/notification-sender/src/notification-sender.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 
 @Injectable()
-export class NotficationSenderService {
+export class NotificationSenderService {
   getHello(): string {
     return 'Hello World!';
   }
