+++ apps/notification-sender/test/jest-e2e.json
